pyproject.toml
[tool.poetry]
name = "tailucas-pylib"
version = "0.0.1"
description = "Common Python utility modules"
authors = ["Tai Lucas <tglucas@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
botoflow = "^0.8"
boto3 = "^1.26.68"
cronitor = "^4.6.0"
msgpack = "^1.0.4"
onepasswordconnectsdk = "^1.3.0"
pika = "^1.3.1"
python-dateutil = "^2.8.2"
pytz = "^2022.7.1"
pyzmq = "^25.0.0"
sentry-sdk = "^1.15.0"
simplejson = "^3.18.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
